Data display methods and moviesScreen view adjustments

+++ TodoMovies/TodoMovies/Models/moviesList.swift
@@ -10,7 +10,7 @@ import Foundation
 struct moviesList: Identifiable, Codable {
     var id: Int
     var listName: String
-    var numberOfLikes: Float
+    var numberOfLikes: Double
     var numberOfViews: Int
     //var listAuthor: String
     var image: String
@@ -23,5 +23,3 @@ struct moviesList: Identifiable, Codable {
         case image = "backdrop_path"
     }
 }
-
-
